pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ielts-speaking-tailor"
version = "0.1.0"
description = "CLI for generating personalized IELTS speaking answer documents from replaceable question banks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "python-docx>=1.1.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
ielts-tailor = "ielts_tailor.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
